Partial fix for part 1 of issue #7.
My current thinking is that the abstract vs concrete distinction of
Structure is more of an implementation detail in Qri, and may not be
worth bringing up in this discussion.  I think the basic point of the
deterministic querying approach can be described without delving into
this.  Might be wrong...